Azure_function_app/requirements.txt
pandas=0.25.1
azure-storage-blob=2.1.0
requests=2.22.0